requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
packaging==24.2
pillow==11.1.0
pydub==0.25.1
pyTelegramBotAPI==4.26.0
pytesseract==0.3.13
requests==2.32.3
SpeechRecognition==3.14.1
typing_extensions==4.12.2
urllib3==2.3.0
